fix+docs: v0.6 全量审计修复 4 个 bug + 补全文档
**背景**：v0.6 的 7 个开发项此前未走「审计存档 + 写文档」环节，本次全量重读文件补做审计 + 修复 + 文档。

**审计发现并修复（4 个代码 bug）**：
1. CRITICAL logging_utils.log_metrics：tensorboard 双前缀（metrics key 已带 loss/ 前缀 + add_scalars main_tag 又加 → loss/loss/total）
2. HIGH schedulers.get_lr/step：step() 直接改 group[lr] 违反 _LRScheduler 基类契约
3. MEDIUM run_ensemble.average_models：.float() 把 float64 降为 float32 精度损失
4. MEDIUM collect_pareto：params 恒为 0（未读 checkpoint n_params）

**补写文档**：
- docs/dev_reference/api.md：CLI 表补 9 个 v0.6 flag + 4.5/4.6/4.7 新模块 API
- CHANGELOG.md：v0.6 段
- docs/architecture/架构设计文档.md：文件结构图登记 6 个新文件
- docs/experiment_reports/调参指南.md：六、P17 loss_prop + 七、P12 bc_loss_type
- docs/audit/2026-08-14-v0.6-审计报告.md：审计存档

**验证**：pytest 12 passed + smoke_test 11 步全通过 + 文档签名与代码 inspect 一致

**影响范围**：修复后 logging_utils tensorboard 标签正确、scheduler 遵循基类契约、集成保留 float64、Pareto 有真实 params

+++ run_ensemble.py
@@ -60,10 +60,10 @@ def average_models(ckpt_paths: list[Path], out_path: str) -> dict:
         state = torch.load(p, map_location="cpu", weights_only=False)
         states.append(state["model_state_dict"])
 
-    # 平均
+    # 平均（保持原 dtype，float64 精度不损失）
     avg_state = {}
     for key in states[0].keys():
-        tensors = torch.stack([s[key].float() for s in states])
+        tensors = torch.stack([s[key] for s in states])
         avg_state[key] = tensors.mean(dim=0)
 
     # 保存 merged（保留第一个 checkpoint 的元数据）
